+++ src/test/app/agent/agent_it.py
@@ -13,7 +13,7 @@ from test.app.test_functions import delete_saved_files, init_logging, load_agent
 from aideas.app.agent.agent_name import AgentName
 from aideas.app.agent.browser_agent import BrowserAgent
 from aideas.app.config import Name, STAGES_KEY, STAGE_ITEMS_KEY
-from aideas.app.result.result_set import ElementResultSet
+from aideas.app.result.result_set import ElementResultSet, StageResultSet
 from aideas.app.run_context import RunContext
 
 from pyu.io.file import load_yaml_str
@@ -40,7 +40,6 @@ class AgentIT(unittest.TestCase):
 # TODO This test works when run alone, but fails when run with the next test. Find out why and fix.
 #
 #     def test_ask_for_help_at_stage(self):
-#         time.sleep(2)
 #         agent_name: str = "test-agent"
 #         run_context: RunContext = get_run_context([agent_name])
 #         yaml = """
@@ -53,14 +52,16 @@ class AgentIT(unittest.TestCase):
 #         agent = Agent.of_config(
 #             agent_name, run_context.get_app_config().to_dict(), agent_config)
 #
+#         result_set = StageResultSet.none()
 #         try:
-#             agent.run(run_context)
+#             result_set = agent.run(run_context)
 #             self.fail("Should raise ActionError because help was not provided in this test case, but did not.")
 #         except ActionError:
 #             logging.info("ActionError raised as expected, test passed.")
+#         finally:
+#             [delete_saved_files(result_set.get(k)) for k in result_set.keys()]
 
     def test_ask_for_help_at_stage_item(self):
-        time.sleep(2)
         agent_name: str = "test-agent"
         run_context: RunContext = get_run_context([agent_name])
         yaml = """
@@ -77,11 +78,14 @@ stages:
         agent = Agent.of_config(
             agent_name, run_context.get_app_config().to_dict(), agent_config)
 
+        result_set = StageResultSet.none()
         try:
-            agent.run(run_context)
+            result_set = agent.run(run_context)
             self.fail("Should raise ActionError because help was not provided in this test case, but did not.")
         except ActionError:
             logging.info("ActionError raised as expected, test passed.")
+        finally:
+            [delete_saved_files(result_set.get(k)) for k in result_set.keys()]
 
     def test_ask_for_help_browser_agent(self):
         agent_name: str = "test-agent"
@@ -100,11 +104,14 @@ stages:
         agent = BrowserAgent.of_config(
             agent_name, run_context.get_app_config().to_dict(), agent_config)
 
+        result_set = StageResultSet.none()
         try:
-            agent.run(run_context)
+            result_set = agent.run(run_context)
             self.fail("Should raise ActionError because help was not provided in this test case, but did not.")
         except ActionError:
             logging.info("ActionError raised as expected, test passed.")
+        finally:
+            [delete_saved_files(result_set.get(k)) for k in result_set.keys()]
 
     def test_pictory_saved_video_is_available_for_other_agents(self):
         run_context: RunContext = get_run_context([pictory_agent])